usb/requirements.txt
pyusb
urllib3
crc
influxdb_client
PyYAML
rich

